Introduce `AbstractJsonTokenParser` class
Move common functions out of the `JsonTokenParser` class that will
be used in the sync and async parsers. They are now placed in
`AbstractJsonTokenParser`. Introduce a helper function in this
abstract class as well to extract data needed to look up verifiers
in both sync and async variants of `VerifierProviders`.
